+++ dune/stuff/fem/customprojection.hh
@@ -199,7 +199,7 @@ protected:
 
 namespace Experimental {
 
-template< int dim, class RangeType, class JacobianRangeType >
+template< size_t dim, class RangeType, class JacobianRangeType >
 struct GradientJacobianToLaplacian
   : public RangeType
 {
